Create README.md
Employee Attrition is where employees leave a company or an organization they were working for.
Employee attrition can be for voluntary or involuntary reasons.
The reasons are through natural means like retirement, or it can be through resignation, termination of contract. 
It costs precious time and money and can result in a loss of staff morale/motivation. This could also tarnish a company's reputation.